duCmd logic movied to du.go
